requirements-old.txt
langchain==0.3.27
langchain-community==0.3.31
langchain_groq==0.3.8
langchain_huggingface==0.3.1
faiss_cpu   #facebook database
pypdf
huggingface_hub
flask
sentence-transformers==5.1.1
python-dotenv
bert-score 
scikit-learn 
matplotlib 
seaborn 
ragas
evaluate
datasets
pandas
